integrate permissions and roles

+++ Domain/Services/CustomerService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Domain.Entities;
+using Domain.Enums;
 using Domain.IRepositories;
 using Domain.IServices;
 using Domain.Models;
@@ -15,21 +16,41 @@ namespace Domain.Services
     public class CustomerService : ICustomerService
     {
         private readonly ICustomerRepository _customerRepository;
- 
-       
-        public CustomerService(ICustomerRepository customerRepository) {
+        private readonly IUserRepository _userRepository;
+
+        public CustomerService(ICustomerRepository customerRepository ,IUserRepository userRepository) {
             this._customerRepository = customerRepository;
+            this._userRepository = userRepository;
+        }
+        public async Task<bool> addCustomer(CustomerVM customer)
+        {
+
+            if (customer != null)
+            {
+                UserVM user = _userRepository.getUserByID(customer.userID).Result;
+                long userId = -1;
+                if (user != null)
+                    userId = user.PermissionID;
+                if (userId == ((int)(UserLookups.MaintainEntities)) + 1)
+                    return await _customerRepository.addCustomerAsync(customer);
+            }
+            return false;
 
         }
-        public Task<bool> addCustomer(CustomerVM customer)
-        {
-            return _customerRepository.addCustomerAsync(customer);
 
-        }
-
-        public Task<bool> deleteCustomerByID(long ID)
+        public async Task<bool> deleteCustomerByID(long ID)
         {
-            return _customerRepository.deleteCustomerByID(ID); 
+            CustomerVM customer = getCustomerByID(ID).Result;
+            if (customer != null)
+            {
+                UserVM user = _userRepository.getUserByID(customer.userID).Result;
+                long userId = -1;
+                if (user != null)
+                    userId = user.PermissionID;
+                if (userId == ((int)(UserLookups.MaintainEntities)) + 1)
+                    return await _customerRepository.deleteCustomerByID(ID);
+            }
+            return false;
         }
 
         public   Task<IEnumerable<CustomerVM>> getAllCustomers()
@@ -43,9 +64,18 @@ namespace Domain.Services
         }
 
   
-        public Task<bool> updateCustomerByID(long ID, CustomerVM customer)
+        public async Task<bool> updateCustomerByID(long ID, CustomerVM customer)
         {
-            return _customerRepository.updateCustomerByID(ID, customer);
+            if (customer != null)
+            {
+                UserVM user = _userRepository.getUserByID(customer.userID).Result;
+                long userId = -1;
+                if (user != null)
+                    userId = user.PermissionID;
+                if (userId == ((int)(UserLookups.MaintainEntities)) + 1)
+                    return await _customerRepository.updateCustomerByID(ID, customer);
+            }
+            return false;
         }
     }
 }
